Move screens to screens directory
